gae/standard_requirements.txt
flask==0.12
werkzeug==0.12.2
google-auth
google-api-python-client
google-auth-httplib2
WebTest
